// rallyregressiondashboard/src/types/rally/test-cases/test-case.ts
import { LastResult } from "../results/last-result";
import { Result } from "../results/result";
import { ResultCount } from "../results/result-count";

import { Milestone } from "../results/result-milestone";

export type TestCase = {
  _rallyAPIMajor: string | null;
  _rallyAPIMinor: string | null;
  _ref: string | null;
  _refObjectUUID: string | null;
  _objectVersion: string | null;
  _refObjectName: string | null;
  CreationDate: string | null;
  _CreatedAt: string | null;
  ObjectID: number | null;
  ObjectUUID: string | null;
  VersionId: string | null;
  Subscription: Result | null;
  Workspace: Result | null;
  Changesets: ResultCount | null;
  Connections: ResultCount | null;
  CreatedBy: Result | null;
  Description: string | null;
  Discussion: ResultCount | null;
  DisplayColor: string | null;
  Expedite: boolean | null;
  FormattedID: string | null;
  LastUpdateDate: string | null;
  LatestDiscussionAgeInMinutes: any | null;
  Milestones: Milestone | null;
  Name: string | null;
  Notes: string | null;
  Owner: Result | null;
  Project: Result | null;
  Ready: boolean | null;
  RevisionHistory: LastResult | null;
  Tags: Milestone | null;
  Attachments: ResultCount | null;
  DefectStatus: string | null;
  Defects: ResultCount | null;
  DragAndDropRank: string | null;
  LastBuild: string | null;
  LastResult: LastResult | null;
  LastRun: string | null;
  LastVerdict: string | null;
  Method: string | null;
  Objective: string | null;
  Package: any | null;
  PostConditions: string | null;
  PreConditions: string | null;
  Priority: string | null;
  Recycled: false | null;
  Results: ResultCount | null;
  Risk: string | null;
  Steps: ResultCount | null;
  TestFolder: Result | null;
  TestSets: ResultCount | null;
  Type: string | null;
  ValidationExpectedResult: string | null;
  ValidationInput: string | null;
  WorkProduct: Result | null;
  c_ALMID: any | null;
  c_APITest: boolean | null;
  c_AssignedTo: any | null;
  c_AutomatedExecution: any | null;
  c_AutomationStatus: any | null;
  c_AutomationTool: any | null;
  c_Component: any | null;
  c_ComponentDRFIS: any | null;
  c_ComponentRetail360: any | null;
  c_DMSInitiatives: any | null;
  c_DMSPriority: any | null;
  c_ManualExecution: any | null;
  c_RTSProduct: Milestone | null;
  c_SubCategoryCAIIS: any | null;
  c_TicketID: any | null;
  c_ToBeAutomated: boolean | null;
  _type: string | null;
};